+++ src/main/java/com/memesphere/domain/user/service/AuthServiceImpl.java
@@ -13,6 +13,7 @@ import com.memesphere.global.redis.RedisService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 
 
@@ -111,4 +112,28 @@ public class AuthServiceImpl implements AuthService{
         return userRepository.findByNickname(nickname).isPresent();
     }
 
+    public String getTmpPassword() {
+        char[] charSet = new char[]{ '0', '1', '2', '3', '4', '5', '6', '7', '8', '9',
+                'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N',
+                'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z'};
+
+        String pwd = "";
+
+        // 문자 배열 길이의 값을 랜덤으로 10개를 뽑아 조합
+        int idx = 0;
+        for(int i = 0; i < 10; i++){
+            idx = (int) (charSet.length * Math.random());
+            pwd += charSet[idx];
+        }
+
+        return pwd;
+    }
+
+    @Transactional
+    public void updatePassword(String tmpPassword, String memberEmail) {
+        String encryptPassword = passwordEncoder.encode(tmpPassword);
+        User existingUser = userRepository.findByEmail(memberEmail).orElseThrow(() -> new GeneralException(ErrorStatus.USER_NOT_FOUND));
+
+        existingUser.updatePassword(encryptPassword);
+    }
 }
