--- src/main/java/com/memesphere/domain/user/service/UserService.java
@@ -1,9 +1,0 @@
-package com.memesphere.domain.user.service;
-
-import com.memesphere.domain.user.entity.User;
-
-public interface UserService {
-    User findByLoginId(Long loginId);
-    public String getTmpPassword();
-    public void updatePassword(String tmpPassword, String memberEmail);
-}
