--- src/main/java/com/memesphere/domain/user/service/UserServiceImpl.java
@@ -1,49 +1,0 @@
-package com.memesphere.domain.user.service;
-
-import com.memesphere.domain.user.entity.User;
-import com.memesphere.domain.user.repository.UserRepository;
-import com.memesphere.global.apipayload.code.status.ErrorStatus;
-import com.memesphere.global.apipayload.exception.GeneralException;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-public class UserServiceImpl implements UserService {
-
-    private final UserRepository userRepository;
-    private final PasswordEncoder passwordEncoder;
-
-    public User findByLoginId(Long loginId) {
-        return userRepository.findByLoginId(loginId).orElse(null);
-    }
-
-    @Override
-    public String getTmpPassword() {
-        char[] charSet = new char[]{ '0', '1', '2', '3', '4', '5', '6', '7', '8', '9',
-                'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N',
-                'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z'};
-
-        String pwd = "";
-
-        /* 문자 배열 길이의 값을 랜덤으로 10개를 뽑아 조합 */
-        int idx = 0;
-        for(int i = 0; i < 10; i++){
-            idx = (int) (charSet.length * Math.random());
-            pwd += charSet[idx];
-        }
-
-        return pwd;
-    }
-
-    @Transactional
-    @Override
-    public void updatePassword(String tmpPassword, String memberEmail) {
-        String encryptPassword = passwordEncoder.encode(tmpPassword);
-        User existingUser = userRepository.findByEmail(memberEmail).orElseThrow(() -> new GeneralException(ErrorStatus.USER_NOT_FOUND));
-
-        existingUser.updatePassword(encryptPassword);
-    }
-}
